+++ src/components/PregnancyComponents/ClinicalConservation.jsx
@@ -1,0 +1,5 @@
+const ClinicalConservation = ({ formObject, setFormObject, handleChange }) => {
+  return <></>;
+};
+
+export default ClinicalConservation;
