+++ src/components/PregnancyComponents/MarkAttendance.jsx
@@ -1,0 +1,5 @@
+const MarkAttendance = ({ formObject, setFormObject, handleChange }) => {
+  return <></>;
+};
+
+export default MarkAttendance;
