+++ src/components/PregnancyComponents/PrenatalCareRelated.jsx
@@ -1,0 +1,5 @@
+const PrenatalCareRelated = ({ formObject, setFormObject, handleChange }) => {
+  return <></>;
+};
+
+export default PrenatalCareRelated;
