+++ src/components/userComponents/dateInput.jsx
@@ -5,7 +5,7 @@ import { DatePicker } from "@mui/x-date-pickers/DatePicker";
 
 const DateInput = ({ title, index, placeholder, value = "", onChange }) => {
   return (
-    <div className="grid grid-cols-3 gap-4 items-center mt-4 mx-14">
+    <div className="grid grid-cols-4 gap-4 items-center mt-4">
       <div className="w-fit">
         {index >= 0 ? (
           `${index + 1}. ${title}`
